show deletion account when ios platform

+++ lib/pages/account/account_page.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:rumah_kreatif_toba/base/confirmation_dialog.dart';
@@ -230,33 +232,33 @@ class AccountPage extends StatelessWidget {
                                 ),
 
                                 //Deletion Account
-                                // if (Platform.isIOS)
-                                GestureDetector(
-                                  onTap: () async {
-                                    showConfirmPopUp(
-                                      title: "Yakin ingin menghapus akun?",
-                                      subTitle:
-                                          "Setelah menghapus akun, seluruh data akan dihapus dari aplikasi dan tidak dapat dipulihkan termasuk riwayat transaksi,foto profile, dll",
-                                      onAccept: () {
-                                        Get.find<UserController>()
-                                            .hapusAkun();
-                                      },
-                                      onCancel: () => Get.back(),
-                                    );
-                                  },
-                                  child: AccountWidget(
-                                      appIcon: AppIcon(
-                                        icon: Icons.delete_forever,
-                                        backgroundColor:
-                                            Colors.white.withOpacity(0.0),
-                                        iconColor: AppColors.redColor,
-                                        iconSize: Dimensions.height10 * 5 / 2,
-                                        size: Dimensions.height10 * 5,
-                                      ),
-                                      bigText: BigText(
-                                        text: "Hapus Akun",
-                                      )),
-                                ),
+                                if (Platform.isIOS)
+                                  GestureDetector(
+                                    onTap: () async {
+                                      showConfirmPopUp(
+                                        title: "Yakin ingin menghapus akun?",
+                                        subTitle:
+                                            "Setelah menghapus akun, seluruh data akan dihapus dari aplikasi dan tidak dapat dipulihkan termasuk riwayat transaksi,foto profile, dll",
+                                        onAccept: () {
+                                          Get.find<UserController>()
+                                              .hapusAkun();
+                                        },
+                                        onCancel: () => Get.back(),
+                                      );
+                                    },
+                                    child: AccountWidget(
+                                        appIcon: AppIcon(
+                                          icon: Icons.delete_forever,
+                                          backgroundColor:
+                                              Colors.white.withOpacity(0.0),
+                                          iconColor: AppColors.redColor,
+                                          iconSize: Dimensions.height10 * 5 / 2,
+                                          size: Dimensions.height10 * 5,
+                                        ),
+                                        bigText: BigText(
+                                          text: "Hapus Akun",
+                                        )),
+                                  ),
                                 SizedBox(
                                   height: Dimensions.height20,
                                 ),
